+++ packages/vue-instantsearch/src/components/Pagination.vue
@@ -30,7 +30,8 @@
   </ul>
 </template>
 
-<script>import algoliaComponent from '../component';
+<script>
+import algoliaComponent from '../component';
 
 export default {
   mixins: [algoliaComponent],