+++ packages/vue-instantsearch/src/components/Rating.vue
@@ -27,7 +27,8 @@
   </div>
 </template>
 
-<script>import { FACET_OR } from '../store';
+<script>
+import { FACET_OR } from '../store';
 import algoliaComponent from '../component';
 
 export default {
@@ -55,7 +56,9 @@ export default {
     this.searchStore.addFacet(this.attributeName, FACET_OR);
   },
   destroyed() {
+    this.searchStore.stop();
     this.searchStore.removeFacet(this.attributeName);
+    this.searchStore.start();
   },
   computed: {
     show() {