+++ packages/vue-instantsearch/src/components/RefinementList.vue
@@ -24,7 +24,8 @@
   </div>
 </template>
 
-<script>import { FACET_OR, FACET_AND } from '../store';
+<script>
+import { FACET_OR, FACET_AND } from '../store';
 import algoliaComponent from '../component';
 
 export default {
@@ -62,7 +63,9 @@ export default {
     this.searchStore.addFacet(this.attributeName, this.operator);
   },
   destroyed() {
+    this.searchStore.stop();
     this.searchStore.removeFacet(this.attributeName);
+    this.searchStore.start();
   },
   computed: {
     facetValues() {