+++ packages/vue-instantsearch/src/components/Results.vue
@@ -12,7 +12,8 @@
   </div>
 </template>
 
-<script>import algoliaComponent from '../component';
+<script>
+import algoliaComponent from '../component';
 
 export default {
   mixins: [algoliaComponent],