+++ packages/vue-instantsearch/src/components/ResultsPerPageSelector.vue
@@ -6,7 +6,8 @@
   </select>
 </template>
 
-<script>import algoliaComponent from '../component';
+<script>
+import algoliaComponent from '../component';
 
 export default {
   mixins: [algoliaComponent],