+++ packages/vue-instantsearch/src/components/SortBySelector.vue
@@ -8,7 +8,8 @@
   </select>
 </template>
 
-<script>import algoliaComponent from '../component';
+<script>
+import algoliaComponent from '../component';
 
 export default {
   mixins: [algoliaComponent],