+++ packages/vue-instantsearch/src/components/__tests__/refinement-list.js
@@ -17,6 +17,8 @@ const addFacet = jest.fn();
 const searchStore = {
   getFacetValues,
   addFacet,
+  stop: () => {},
+  start: () => {},
 };
 
 test('renders proper HTML', () => {