+++ packages/vue-instantsearch/stories/Clear.stories.js
@@ -5,4 +5,11 @@ storiesOf('Clear', module)
   .addDecorator(previewWrapper)
   .add('default', () => ({
     template: '<ais-clear></ais-clear>',
+  }))
+  .add('custom rendering', () => ({
+    template: `<ais-clear>
+      <template>
+        Clear search query
+      </template>
+    </ais-clear>`,
   }));