+++ packages/vue-instantsearch/stories/NoResults.stories.js
@@ -8,4 +8,12 @@ storiesOf('NoResults', module)
     mounted() {
       this.$refs.child.searchStore.query = 'no results for this';
     },
+  }))
+  .add('custom rendering', () => ({
+    template: `<ais-no-results ref="child">
+      <span slot-scope="{query}">no results for "{{query}}"</span>
+    </ais-no-results>`,
+    mounted() {
+      this.$refs.child.searchStore.query = 'no results for this';
+    },
   }));