+++ packages/vue-instantsearch/CHANGELOG.md
@@ -1,3 +1,28 @@
+<a name="1.3.2"></a>
+## [1.3.2](https://github.com/algolia/vue-instantsearch/compare/v1.3.1...v1.3.2) (2017-11-15)
+
+
+### Bug Fixes
+
+* avoid query when components are destroyed ([c41c606](https://github.com/algolia/vue-instantsearch/commit/c41c606))
+
+
+
+<a name="1.3.1"></a>
+## [1.3.1](https://github.com/algolia/vue-instantsearch/compare/v1.3.0...v1.3.1) (2017-11-02)
+
+
+### Bug Fixes
+
+* **menu:** avoid multiple Algolia calls when created ([f1ac0a7](https://github.com/algolia/vue-instantsearch/commit/f1ac0a7))
+
+
+### Performance Improvements
+
+* **menu:** use facet name as key ([b496fd8](https://github.com/algolia/vue-instantsearch/commit/b496fd8))
+
+
+
 <a name="1.3.0"></a>
 # [1.3.0](https://github.com/algolia/vue-instantsearch/compare/v1.2.2...v1.3.0) (2017-10-25)